Gutting django-registration for #37
This basically breaks everything, since nothing knows where to go to
require login, and everything requires login. However, it's a necessary
step in replacing django-registration. Needless to say, stuff won't work
in this branch for quite a while, and there's a lot of work ahead of
us. First off is getting a basic auth system in place.