Add image upload as part of #21
Committing stuff mostly from Friday morning for EOH, this uploads images
but does not display them anywhere yet.

Full file upload (being able to upload any file, and having the site
detect and display images seperately) is a goal for the future, but it
is too complex to implement now.